clean up metal code
