+++ llama/llama.go
@@ -37,8 +37,6 @@ package llama
 #include "sampling_ext.h"
 
 bool llamaProgressCallback(float progress, void *user_data);
-extern const char *ggml_metallib_start;
-extern const char *ggml_metallib_end;
 */
 import "C"
 import (
@@ -50,19 +48,6 @@ import (
 	"strings"
 	"unsafe"
 )
-
-//go:embed ggml-common.h
-var ggmlCommon string
-
-//go:embed ggml-metal.metal
-var ggmlMetal string
-
-func init() {
-	metal := strings.ReplaceAll(ggmlMetal, `#include "ggml-common.h"`, ggmlCommon)
-	cMetal := C.CString(metal)
-	C.ggml_metallib_start = cMetal
-	C.ggml_metallib_end = (*C.char)(unsafe.Pointer(uintptr(unsafe.Pointer(cMetal)) + uintptr(len(metal))))
-}
 
 func BackendInit() {
 	C.llama_backend_init()
